mini_project_2/requirements.txt
chromadb
langchain
sentence-transformers
ollama
pypdf2
python-docx
